fixed dirty-chai to enable test debugging, moved `logger.js` to `util/logger.js` and renamed `http_services` to `server.js`. And some small linting fixes.
