+++ bin/server.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const config = require('../src/config');
-const httpServices = require('../src/http_services');
-const logger = require('../src/logger');
+const httpServices = require('../src/server');
+const logger = require('../src/util/logger');
 
 httpServices
   .start(config.app.port)
