+++ tests/acceptance/step_definitions/authors/the_last_response_body_contains_the_authors_data.js
@@ -6,7 +6,7 @@ module.exports = function() {
   this.Then(/^the last response body contains the authors data$/, done => {
     const world = this.world;
     const authors = world.lastResponseBody.authors;
-    
+
     authors.forEach(author => expect(world.authorsToCreate).to.include(author));
     done();
   });
