+++ tests/integration/dirty-chai.js
@@ -4,3 +4,5 @@ const chai = require('chai');
 const dirtyChai = require('dirty-chai');
 
 chai.use(dirtyChai);
+
+module.exports = chai;
