+++ tests/integration/storage/connect_test.js
@@ -1,16 +1,15 @@
 'use strict';
 
 const connect = require('../../../src/storage/connect');
-const expect = require('chai').expect;
+const expect = require('../dirty-chai').expect;
 
 describe('Storage connect integration tests', () => {
 
   it('Should connect to mongodb', done => {
     connect().then(db => {
       expect(db).to.exist();
-      done();
-
-    }).catch(done.reject);
+      return done();
+    }).catch(done);
   });
 
 });
